// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ContestStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  ENDED: 'ENDED'
} as const

export type ContestStatus = (typeof ContestStatus)[keyof typeof ContestStatus]


export const ContestType = {
  PRIVATE: 'PRIVATE',
  PUBLIC: 'PUBLIC'
} as const

export type ContestType = (typeof ContestType)[keyof typeof ContestType]


export const ContestRole = {
  TEAM: 'TEAM',
  JUDGE: 'JUDGE',
  BALLOON: 'BALLOON',
  OBSERVER: 'OBSERVER',
  ADMIN: 'ADMIN'
} as const

export type ContestRole = (typeof ContestRole)[keyof typeof ContestRole]


export const Verdict = {
  PENDING: 'PENDING',
  JUDGING: 'JUDGING',
  ACCEPTED: 'ACCEPTED',
  WRONG_ANSWER: 'WRONG_ANSWER',
  TIME_LIMIT_EXCEEDED: 'TIME_LIMIT_EXCEEDED',
  MEMORY_LIMIT_EXCEEDED: 'MEMORY_LIMIT_EXCEEDED',
  RUNTIME_ERROR: 'RUNTIME_ERROR',
  COMPILE_ERROR: 'COMPILE_ERROR',
  PRESENTATION_ERROR: 'PRESENTATION_ERROR',
  SYSTEM_ERROR: 'SYSTEM_ERROR'
} as const

export type Verdict = (typeof Verdict)[keyof typeof Verdict]


export const ClariCategory = {
  QUESTION: 'QUESTION',
  NOTICE: 'NOTICE'
} as const

export type ClariCategory = (typeof ClariCategory)[keyof typeof ClariCategory]